Make position and duration integer properties.

+++ src/multimedia/qmlaudio_p.h
@@ -57,8 +57,8 @@ class QmlAudio : public QObject, public QmlMediaBase, public QmlParserStatus
     Q_PROPERTY(QUrl source READ source WRITE setSource NOTIFY sourceChanged)
     Q_PROPERTY(QmlMedia::State state READ state WRITE setState NOTIFY stateChanged)
     Q_PROPERTY(QmlMedia::Status status READ status NOTIFY statusChanged)
-    Q_PROPERTY(qreal duration READ duration NOTIFY durationChanged)
-    Q_PROPERTY(qreal position READ position WRITE setPosition NOTIFY positionChanged)
+    Q_PROPERTY(int duration READ duration NOTIFY durationChanged)
+    Q_PROPERTY(int position READ position WRITE setPosition NOTIFY positionChanged)
     Q_PROPERTY(qreal volume READ volume WRITE setVolume NOTIFY volumeChanged)
     Q_PROPERTY(bool muted READ isMuted WRITE setMuted NOTIFY mutedChanged)
     Q_PROPERTY(int bufferStatus READ bufferStatus NOTIFY bufferStatusChanged)
@@ -76,8 +76,8 @@ Q_SIGNALS:
     void stateChanged(QmlMedia::State state);
     void statusChanged(QmlMedia::Status status);
 
-    void durationChanged(qreal duration);
-    void positionChanged(qreal position);
+    void durationChanged(int duration);
+    void positionChanged(int position);
 
     void volumeChanged(qreal volume);
     void mutedChanged(bool muted);
